requirements.txt
Django==5.12
pytesseract==0.3.13
Pillow==11.0.0
requests==2.32.3
PyMuPDF==1.24.12
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.3
openai==1.52.0
pytest-shutil==1.8.0
cohere==5.11.1

